# static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AI Interview Platform API",
    "description": "API documentation for the AI-Powered Technical Interview Platform",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api",
      "description": "API Server"
    }
  ],
...
  "paths": {
...
    "/interviews": {
      "get": {
        "tags": ["Interviews"],
        "summary": "Get user's interviews",
        "description": "Get interviews for the authenticated user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "description": "Filter by status (scheduled, completed, cancelled)",
            "schema": {
              "type": "string",
              "enum": ["scheduled", "completed", "cancelled"]
            }
          },
          {
            "name": "from_date",
            "in": "query",
            "description": "Filter by date range start (YYYY-MM-DD)",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "to_date",
            "in": "query",
            "description": "Filter by date range end (YYYY-MM-DD)",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of interviews",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Interviews"],
        "summary": "Schedule a new interview",
        "description": "Create a new interview (recruiters only)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["candidate_id", "position_id", "scheduled_time"],
                "properties": {
                  "candidate_id": {
                    "type": "integer"
                  },
                  "position_id": {
                    "type": "integer"
                  },
                  "scheduled_time": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "duration_minutes": {
                    "type": "integer",
                    "default": 60
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Interview scheduled successfully"
          },
          "400": {
            "description": "Invalid request"
          },
          "403": {
            "description": "Unauthorized access"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}